Add helpful message when TensorBoard manager is not available
Added informative message when tensorboard_manager is None, which happens
if the user is in an arc chat session that was started before the
TensorBoard feature was added.

Message: 'TensorBoard auto-launch not available (restart arc chat to enable)'

This helps users understand why they're seeing manual instructions instead
of the interactive TensorBoard preference dialog.